Add FLOAT type basic implementation
The type is not yet fully supported by server.

Basic functionality covers:
* reading and writing to BlockBuilder
* literal constructor

This commit will allow to continue work on providing floating
point type smaller (4B) than currently available DOUBLE.

It also makes MaterializedRow float precision agnostic

Before FLOAT was implemented MaterializedRow was converting single-precision floats to doubles.
This caused failures on results comparison while float came to life.

+++ presto-main/src/test/java/com/facebook/presto/type/TestArrayOperators.java
@@ -93,7 +93,7 @@ public class TestArrayOperators
     public void testTypeConstructor()
             throws Exception
     {
-        assertFunction("ARRAY[7]", new ArrayType(BIGINT), ImmutableList.of(7L));
+        assertFunction("ARRAY[7]", new ArrayType(INTEGER), ImmutableList.of(7));
         assertFunction("ARRAY[12.34, 56.78]", new ArrayType(DOUBLE), ImmutableList.of(12.34, 56.78));
     }
